+++ packages/integration-tests/src/integrationTests/assertions.ts
@@ -1,4 +1,4 @@
-import { ErrorPayload, publicChannels, SocketActionTypes, TestStore, messages, MessageType } from '@quiet/state-manager'
+import { ErrorPayload, publicChannels, TestStore, messages, MessageType } from '@quiet/state-manager'
 import waitForExpect from 'wait-for-expect'
 import { MAIN_CHANNEL } from '../testUtils/constants'
 import { sleep } from '../utils'